Fixing critical issue in removeMethodCall
Issue: if loadTaskWithId pulls the task from the database (not the
cache) and if that task is enqueued or running, then
getTaskWithCompletionHandler may not have that task as part of its list
of downloadTasks (because it may have been started a long time ago
and gotten lost as the app was killed).
In that case, the taskId will not be found in the list, and therefore
the task will not be removed from the database.  As a result, the\
sqlite database can get polluted with tasks that cannot be removed
ever.

Fix: call [self deleteTask:taskId] unconditionally after the
if-statement (that checks for enqueue or running) instead of only
in the 'else' clause of that if-statement